scripts/requirements.txt
# Python CLI tools for inventory management
# Install: pip install -r scripts/requirements.txt

# Supabase client
supabase>=2.3.0

# Environment variables
python-dotenv>=1.0.0
